Reorganize based off raster utils vs plane slicer utils
